+++ packages/editor-skeleton/src/transducers/parse-props.ts
@@ -155,6 +155,7 @@ function propTypeToSetter(propType: PropType): SetterType {
   return {
     componentName: 'MixedSetter',
     isRequired,
+    props: {},
   };
 }