+++ packages/rax-simulator-renderer/src/obx-rax/utils/has-own-property.ts
@@ -1,0 +1,5 @@
+export const prototypeHasOwnProperty = Object.prototype.hasOwnProperty;
+
+export function hasOwnProperty(obj: any, key: string | number | symbol): boolean {
+  return obj && prototypeHasOwnProperty.call(obj, key);
+}